Add info on spin timeout
