+++ features/Context/Spin/TimeoutException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Context\Spin;
+
+class TimeoutException extends \Exception
+{
+}
